feat: Add discord upgrade helper for Gentoo
Adds discord upgrade helper for Gentoo to eliminate the hassle of
checking if Gentoo package maintainers have updated the version
available to portage, or whether we should use the portable
executable in the meantime. While it's good to be polite by not
syncing unnecessarily, this is mainly the fault of forced updates
but that's a conversation for a later time.