fix: Improve parsing logic
Improves parsing match patterns to avoid matching non/keys if key
strings appear elsewhere in the table string. For example, a keyname
of 'f' may match the first occurance of the letter 'f' rather
than the intended key. This is fixed by requiring the key to be
directly followed by an '=' sign. Adds optional matching of
whitespace around the '=' sign to avoid parsing errors when no
padding is provided.